requirements.txt
requests==2.31.0
typer>=0.15.1
python-dotenv==1.0.1
